+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: todina-r <todina-r@student.42antananarivo  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/22 08:43:26 by todina-r          #+#    #+#             */
-/*   Updated: 2026/01/22 14:31:03 by todina-r         ###   ########.fr       */
+/*   Updated: 2026/01/22 15:17:59 by todina-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,16 @@
 //#include <string.h>
 //#include <bsd/string.h>
 //#include <stdio.h>
+#include "./libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, int size)
+int	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	int	index;
 	int	src_len;
 
 	index = 0;
 	src_len = ft_strlen(src);
-	while (src[index] && index < (size - 1))
+	while (src[index] && index < (int)(size - 1))
 	{
 		dst[index] = src[index];
 		index++;
