+++ libft.h
@@ -6,12 +6,14 @@
 /*   By: todina-r <todina-r@student.42antananarivo  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/22 14:35:05 by todina-r          #+#    #+#             */
-/*   Updated: 2026/01/22 14:35:31 by todina-r         ###   ########.fr       */
+/*   Updated: 2026/01/22 15:15:17 by todina-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
+
+#include <stdlib.h>
 
 int		ft_isalnum(int c);
 int		ft_isalpha(int c);
@@ -20,17 +22,17 @@ int		ft_isdigit(int c);
 int		ft_isprint(int c);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
-void	ft_bzero(void *s, int n);
+void	ft_bzero(void *s, size_t n);
 void	*ft_memchr(const void *s, int c, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 int		ft_atoi(const char *nptr);
-void	*ft_calloc(size_t nmemb, int size);
+void	*ft_calloc(size_t nmemb, size_t size);
 int		ft_strlen(const char *s);
-int		ft_strncmp(const char *s1, const char *s2, int n);
-int		ft_strlcpy(char *dst, const char *src, int size);
-int		ft_strlcat(char *dst, const char *src, int size);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
+int		ft_strlcpy(char *dst, const char *src, size_t size);
+int		ft_strlcat(char *dst, const char *src, size_t size);
 char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
